home controller
home controller added and changes in article controller
