+++ routes/homeRoutes.js
@@ -1,12 +1,14 @@
 const express = require("express");
 const router = express.Router();
-const articleController = require("../controllers/articleController");
 const commentsController = require("../controllers/commentsController");
+const homeController = require("../controllers/homeController");
 
-router.get("/", articleController.index);
+router.get("/", homeController.index);
 
-router.get("/article/:id", articleController.index);
+router.get("/article/:id", homeController.articles);
 
 router.post("/article/:id/create-comment", commentsController.store);
+
+router.get("/api", homeController.api);
 
 module.exports = router;
